Improvement | Completed refactoring
