+++ modules/gamePlay.js
@@ -1,10 +1,22 @@
-import { loadResultAnnouncement, loadWinnerAnnouncement } from "./utils.js";
+import {
+  loadResultAnnouncement,
+  loadWinnerAnnouncement,
+  options,
+} from "./utils.js";
 
-let playerScore,
+let playerScore = 0,
   comScore = 0;
+console.log(`Player: ${playerScore}`);
+console.log(`Com: ${comScore}`);
+const computerPlay = () => {
+  let int = Math.floor(Math.random() * 3);
+  let optionsArr = Object.keys(options);
+  return optionsArr[int];
+};
 
-const playReverseMode = (playerChoice, comChoice, options) => {
+const playReverseMode = (playerChoice) => {
   let result;
+  let comChoice = computerPlay(options);
   if (playerChoice == comChoice) {
     result = evaluateGamePlay("Draw");
   } else {
@@ -17,33 +29,43 @@ const playReverseMode = (playerChoice, comChoice, options) => {
   return result;
 };
 
-const playNormalMode = (playerChoice, comChoice, options) => {
-  let result;
+const playNormalMode = (playerChoice) => {
+  let comChoice = computerPlay(options);
   if (playerChoice == comChoice) {
-    result = evaluateGamePlay("Draw");
+    console.log(`Player: ${playerScore}`);
+    console.log(`Com: ${comScore}`);
+    loadResultAnnouncement(
+      "Draw",
+      playerChoice,
+      comChoice,
+      playerScore,
+      comScore
+    );
   } else {
     if (options[playerChoice] == comChoice) {
-      result = evaluateGamePlay("Player");
+      playerScore++;
+      console.log(`Player: ${playerScore}`);
+      console.log(`Com: ${comScore}`);
+      loadResultAnnouncement(
+        "Player",
+        playerChoice,
+        comChoice,
+        playerScore,
+        comScore
+      );
     } else {
-      result = evaluateGamePlay("Com");
+      comScore++;
+      console.log(`Player: ${playerScore}`);
+      console.log(`Com: ${comScore}`);
+      loadResultAnnouncement(
+        "Com",
+        playerChoice,
+        comChoice,
+        playerScore,
+        comScore
+      );
     }
   }
-  return result;
-};
-
-const evaluateGamePlay = (result) => {
-  if (result == "Player") {
-    playerScore++;
-  } else if (result == "Com") {
-    comScore++;
-  }
-  loadResultAnnouncement(
-    result,
-    playerChoice,
-    comChoice,
-    playerScore,
-    comScore
-  );
 };
 
 const evaluateWinner = () => {
